requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scipy>=1.11.0
pytest>=7.4.0
pytest-cov>=4.1.0
